pyproject.toml
[project]
name = "realtimestt-onnx"
version = "0.1.0"
description = "A realtime speech-to-text library using ONNX models with advanced voice activity detection"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = [
    "real-time",
    "audio",
    "transcription",
    "speech-to-text",
    "voice-activity-detection",
    "VAD",
    "onnx",
    "speech-recognition",
    "voice-assistants",
    "audio-processing"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "PyAudioWPatch>=0.2.12.6",
    "onnx-asr[cpu,hub]>=0.1.0",
    "webrtcvad-wheels>=2.0.14",
    "numpy>=1.21.6",
    "scipy>=1.7.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "halo>=0.0.31",
]

[project.optional-dependencies]
gpu = [
    "onnx-asr[gpu,hub]>=0.1.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/realtimestt-onnx"
Documentation = "https://github.com/yourusername/realtimestt-onnx#readme"
Repository = "https://github.com/yourusername/realtimestt-onnx"
Issues = "https://github.com/yourusername/realtimestt-onnx/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["realtimestt_onnx"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = []
